Воод переменных от пользователя
