+++ src/main/java/Main.java
@@ -1,26 +1,14 @@
 import java.util.Scanner;
 public class Main {
     public static void main(String[] args) {
-        System.out.println("Введите количество человек:");
-        System.out.println("итого " + amountPersons() + " человек");
+
+        System.out.println(Calculator.amountPersons() + " человек");
+        System.out.println("товар " + Calculator.menu());
+        System.out.println("цена " + Calculator.price());
+
+
     }
 
-    public static int amountPersons() {
-        Scanner scanner = new Scanner(System.in);
-        while (true) {
 
-            if (scanner.hasNextInt()) {
-                int persons = scanner.nextInt();
-                if (persons > 1) {
-                    return persons;
-                } else if (persons <= 1) {
-                        System.out.println("меньше 1");
-                    }
-                } else {
-                System.out.println("не число");
-                scanner.next();}
-                }
-
-            }
         }
